New 'configuration' sub-command with 'valiate' sub-command

See merge request https://gitlab.com/gitlab-org/gitaly/-/merge_requests/5411

Merged-by: Sami Hiltunen <shiltunen@gitlab.com>
Approved-by: Sami Hiltunen <shiltunen@gitlab.com>

+++ internal/gitlab/http_client.go
@@ -52,7 +52,7 @@ func NewHTTPClient(
 		gitlabCfg.RelativeURLRoot,
 		gitlabCfg.HTTPSettings.CAFile,
 		gitlabCfg.HTTPSettings.CAPath,
-		uint64(gitlabCfg.HTTPSettings.ReadTimeout),
+		gitlabCfg.HTTPSettings.ReadTimeout,
 		opts,
 	)
 	if err != nil {
